backend/requirements.txt
fastapi
uvicorn
python-dotenv
mcp
chromadb
google-generativeai
rank_bm25
jieba
numpy